Started working on sleeping the pet
The idea is that once the pet starts sleeping, it will sleep until you click the pet to wake it up (make sure to inform the player of this the first time the sleep the pet). Eventually add an energy and/or time tracker, or some visible way to determine when you should wake the pet. Also, sleep time should take a reasonable amount of time, not be something that goes by to quickly. Like, you put the pet to sleep, do something else, then go back and wake the pet. Maybe in the future there could be an alarm clock item?