Pet Display System
Discarded the sleeping system I built so far cuz I couldn't remember what I was trying to do with it and it was reaally not working. Instead, started trying to implement a basic foundational pet display system which I can later build on for the sleeping pet, as well as other potential future thingiiies.